requirements.txt
awswrangler>=3.2.1
fastapi==0.98.0
mangum==0.17.0
pydantic==1.10.13
google-api-python-client>=2.111.0
google-auth>=2.25.2
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
psycopg2-binary>=2.9.9
sqlalchemy
openpyxl
bs4
cryptography
O365==2.0.38
dateparser
html2text
email_validator
openai
python-dotenv
uvicorn